dom/html/test/test_bug598643.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=598643
-->
<head>
  <title>Test for Bug 598643</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=598643">Mozilla Bug 598643</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 598643 **/

function createFileWithData(fileName, fileData)
{
  return new File([new Blob([fileData], { type: "text/plain" })], fileName);
}

function testFileControl(aElement)
{
  aElement.type = 'file';

  var file = createFileWithData("file_bug598643", "file content");
  SpecialPowers.wrap(aElement).mozSetFileArray([file]);

  ok(aElement.validity.valid, "the file control should be valid");
  ok(!aElement.validity.tooLong,
     "the file control shouldn't suffer from being too long");
}

var types = [
  // These types can be too long.
  [ "text", "email", "password", "url", "search", "tel" ],
  // These types can't be too long.
  [ "radio", "checkbox", "submit", "button", "reset", "image", "hidden",
    'number', 'range', 'date', 'time', 'color', 'month', 'week' ],
  // These types can't be too long but are not implemented yet.
  [ "datetime", 'datetime-local' ]
];

var input = document.createElement("input");
input.maxLength = 1;
input.value = "foo";

// Too long types.
for (type of types[0]) {
  input.type = type
  if (type == 'email') {
    input.value = "foo@bar.com";
  } else if (type == 'url') {
    input.value = 'http://foo.org';
  }

  todo(!input.validity.valid, "the element should be invalid [type=" + type + "]");
  todo(input.validity.tooLong,
       "the element should suffer from being too long [type=" + type + "]");

  if (type == 'email' || type == 'url') {
    input.value = 'foo';
  }
}

// Not too long types.
for (type of types[1]) {
  input.type = type
  ok(input.validity.valid, "the element should be valid [type=" + type + "]");
  ok(!input.validity.tooLong,
     "the element shouldn't suffer from being too long [type=" + type + "]");
}

// Not too long types but TODO.
for (type of types[2]) {
  input.type = type
  todo_is(input.type, type, type + " should not be implemented");
}

testFileControl(input);

</script>
</pre>
</body>
</html>
